+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [7.1.1] - 2019-XX-XX
+### Added
+- Moved High Definition templates and includes to com.unity.render-pipelines.high-definition package
 
 ## [7.0.1] - 2019-07-25
 ### Added